app/pyproject.toml
[tool.poetry]
name = "face-recognition"
version = "0.1.0"
description = ""
authors = ["Babur Parakhatov <melomanovmeloman@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
numpy = ">=1.16.2,<=1.26.0"
opencv-python = ">4,<=4.8.0.76"
multipledispatch = "*"
pype-fastapi = "^0.3.0"
uvicorn = "^0.32.0"
motor = "^3.6.0"
pydantic-settings = "^2.6.1"
onnxruntime = "^1.20.0"
aiofiles = "^24.1.0"
python-multipart = "^0.0.17"
black = "^24.10.0"
gunicorn = "^23.0.0"
redis = "^5.2.0"
aioboto3 = "^13.2.0"
aioredlock = "^0.7.3"
setuptools = "^75.6.0"
httpx = "^0.28.1"
face-sdk-3divi = {path = "python_api"}

[build-system]
requires = ["poetry-core", "setuptools>=75.0.0"]
build-backend = "poetry.core.masonry.api"
